DjangoProject/requirements.txt
dj-database-url==0.5.0
Django==2.2.7
django-filter==2.2.0
djangorestframework==3.10.3
gunicorn==20.0.0
Markdown==3.1.1
Pillow==6.2.1
pytz==2019.3
sqlparse==0.3.0
whitenoise==4.1.4
psycopg2-binary==2.8.3